Replace `float` with `Decimal`
We should use fixed-decimal for precise representation of fractional quantities
such as dollar amounts rather than floating point values which can be imprecise.

+++ interactive_brokers.py
@@ -22,6 +22,7 @@ Does not handle:
 
 import csv
 from datetime import datetime
+from decimal import Decimal
 import utils
 
 
@@ -56,7 +57,7 @@ class InteractiveBrokers:
 
     @classmethod
     def ParseDollarValue(cls, value):
-        return float(value.replace(',', '').replace('"', ''))
+        return Decimal(value.replace(',', '').replace('"', ''))
 
     @classmethod
     def isFileForBroker(cls, filename):
